# docs/openapi/swagger.yaml
openapi: 3.0.0
info:
  title: Social Networking Platform API
  version: 1.1.0
  description: |
    API Gateway contract for the Social Networking Platform.
    Authentication is handled through Google OAuth2, JWTs, and Redis-backed sessions.
servers:
  - url: http://localhost:8080
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    SuccessEnvelope:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
        message:
          type: string
        request_id:
          type: string
    ErrorEnvelope:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              nullable: true
        request_id:
          type: string
...
    Post:
      type: object
      properties:
        id:
          type: string
        authorId:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
    PostInteraction:
      type: object
      properties:
        postId:
          type: string
        postAuthorId:
          type: string
        actorId:
          type: string
        interactionType:
          type: string
          example: like
        createdAt:
          type: integer
          format: int64
...
paths:
...
  /api/v1/posts:
    post:
      tags: [Posts]
      summary: Create a post
      security:
        - bearerAuth: []
      responses:
        '201':
          description: Post created
  /api/v1/posts/{id}:
    get:
      tags: [Posts]
      summary: Get a single post
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Post details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
    patch:
      tags: [Posts]
      summary: Update a post
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Post updated
    delete:
      tags: [Posts]
      summary: Delete a post
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Post deleted
  /api/v1/posts/{id}/interactions:
    post:
      tags: [Posts]
      summary: Interact with a post
      description: Creates a supported post interaction event. Current supported interaction type is `like`.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: X-User-ID
          in: header
          required: true
          schema:
            type: string
          description: Auth subject forwarded by gateway.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                interaction_type:
                  type: string
                  example: like
      responses:
        '202':
          description: Interaction accepted and event publish attempted
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PostInteraction'
        '400':
          description: Invalid post id or unsupported interaction type
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Missing authenticated user identity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Post authors cannot interact with their own post
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Post not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...